+++ uikit/src/main/java/com/sendbird/uikit/internal/ui/notifications/NotificationStatusComponent.kt
@@ -5,6 +5,9 @@ import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import com.sendbird.android.channel.BaseChannel
+import com.sendbird.android.channel.FeedChannel
+import com.sendbird.uikit.R
 import com.sendbird.uikit.internal.model.notifications.NotificationConfig
 import com.sendbird.uikit.modules.components.StatusComponent
 
@@ -22,5 +25,23 @@ internal class NotificationStatusComponent(
             }
         }
         return layout
+    }
+
+    /**
+     * Handles a new channel when data has changed.
+     *
+     * @param channel The latest group channel
+     * since 3.8.0
+     */
+    fun notifyChannelChanged(channel: BaseChannel) {
+        if (channel is FeedChannel && channel.isCategoryFilterEnabled) {
+            rootView?.let {
+                val frameLayout: View = it
+                val layoutParams = frameLayout.layoutParams as ViewGroup.MarginLayoutParams
+                val marginTop = it.resources.getDimensionPixelSize(R.dimen.sb_size_62)
+                layoutParams.topMargin = marginTop
+                frameLayout.layoutParams = layoutParams
+            }
+        }
     }
 }