+++ uikit/src/main/java/com/sendbird/uikit/internal/ui/widgets/NotificationRecyclerView.kt
@@ -1,13 +1,18 @@
 package com.sendbird.uikit.internal.ui.widgets
 
 import android.content.Context
+import android.content.res.ColorStateList
 import android.graphics.drawable.Drawable
 import android.util.AttributeSet
 import android.util.TypedValue
 import android.view.LayoutInflater
 import android.view.MotionEvent
 import android.view.View
+import android.view.ViewGroup
+import android.widget.CompoundButton
 import android.widget.FrameLayout
+import android.widget.RadioButton
+import android.widget.RadioGroup
 import androidx.annotation.ColorInt
 import androidx.constraintlayout.widget.ConstraintSet
 import com.sendbird.uikit.R
@@ -23,11 +28,16 @@ internal class NotificationRecyclerView @JvmOverloads constructor(
     attrs: AttributeSet? = null,
     defStyle: Int = 0
 ) : FrameLayout(context, attrs, defStyle) {
+    private var categoryMenuTextColor: Int
+    private var categoryMenuTextAppearance: Int
+    private var categoryMenuBackground: Int
     private val binding: SbViewChatNotificationRecyclerViewBinding
     val recyclerView: PagerRecyclerView
         get() = binding.rvMessageList
     val isReverseLayout
         get() = recyclerView.isReverseLayout()
+    val categoryFilterBox: RadioGroup
+        get() = binding.categoryMenuBox
 
     override fun setBackground(background: Drawable?) {
         recyclerView.background = background
@@ -63,6 +73,32 @@ internal class NotificationRecyclerView @JvmOverloads constructor(
         binding.vgTooltipBox.setOnClickListener(onTooltipClickListener)
     }
 
+    fun enableCategoryFilterView(enable: Boolean) {
+        binding.svCategoryBox.visibility = if (enable) VISIBLE else GONE
+    }
+
+    fun createCategoryFilterItemView(): CompoundButton {
+        return RadioButton(context).apply {
+            setTextColor(ColorStateList.valueOf(categoryMenuTextColor))
+            setAppearance(context, categoryMenuTextAppearance)
+            setBackgroundResource(categoryMenuBackground)
+            buttonDrawable = null
+            setPadding(
+                resources.intToDp(12),
+                resources.intToDp(7),
+                resources.intToDp(12),
+                resources.intToDp(7)
+            )
+
+            layoutParams = RadioGroup.LayoutParams(
+                ViewGroup.LayoutParams.WRAP_CONTENT,
+                resources.getDimensionPixelSize(com.sendbird.uikit.R.dimen.sb_size_30)
+            ).apply {
+                marginEnd = resources.intToDp(8)
+            }
+        }
+    }
+
     private fun movePosition() {
         val reverse = recyclerView.isReverseLayout()
         val set = ConstraintSet()
@@ -70,18 +106,37 @@ internal class NotificationRecyclerView @JvmOverloads constructor(
         set.clone(rootView)
         if (reverse) {
             set.clear(binding.vgTooltipBox.id, ConstraintSet.TOP)
-            set.connect(binding.vgTooltipBox.id, ConstraintSet.BOTTOM, ConstraintSet.PARENT_ID, ConstraintSet.BOTTOM)
+            set.connect(
+                binding.vgTooltipBox.id,
+                ConstraintSet.BOTTOM,
+                ConstraintSet.PARENT_ID,
+                ConstraintSet.BOTTOM
+            )
         } else {
             set.clear(binding.vgTooltipBox.id, ConstraintSet.BOTTOM)
-            set.connect(binding.vgTooltipBox.id, ConstraintSet.TOP, ConstraintSet.PARENT_ID, ConstraintSet.TOP)
+            set.connect(
+                binding.vgTooltipBox.id,
+                ConstraintSet.TOP,
+                ConstraintSet.PARENT_ID,
+                ConstraintSet.TOP
+            )
         }
         set.applyTo(rootView)
     }
 
     init {
-        val a = context.theme.obtainStyledAttributes(attrs, R.styleable.NotificationListView, defStyle, 0)
+        val a = context.theme.obtainStyledAttributes(
+            attrs,
+            R.styleable.NotificationListView,
+            defStyle,
+            0
+        )
         try {
-            binding = SbViewChatNotificationRecyclerViewBinding.inflate(LayoutInflater.from(getContext()), this, true)
+            binding = SbViewChatNotificationRecyclerViewBinding.inflate(
+                LayoutInflater.from(getContext()),
+                this,
+                true
+            )
             val recyclerViewBackground =
                 a.getResourceId(
                     R.styleable.NotificationListView_sb_notification_recyclerview_background,
@@ -95,12 +150,26 @@ internal class NotificationRecyclerView @JvmOverloads constructor(
                 R.styleable.NotificationListView_sb_notification_recyclerview_tooltip_text_appearance,
                 R.style.SendbirdBody2OnDark01
             )
+
+            categoryMenuBackground = a.getResourceId(
+                R.styleable.NotificationListView_sb_category_filter_background,
+                R.drawable.selector_category_filter_menu_light
+            )
+            categoryMenuTextAppearance = a.getResourceId(
+                R.styleable.NotificationListView_sb_category_filter_text_appearance,
+                R.style.SendbirdCaption2OnLight01
+            )
+            categoryMenuTextColor = a.getResourceId(
+                R.styleable.NotificationListView_sb_category_filter_text_color,
+                R.drawable.selector_category_filter_menu_text_color_light
+            )
             setBackgroundResource(android.R.color.transparent)
             binding.rvMessageList.setOnTouchListener { v: View, _: MotionEvent? ->
                 SoftInputUtils.hideSoftKeyboard(this)
                 v.performClick()
                 false
             }
+            binding.svCategoryBox.setBackgroundResource(recyclerViewBackground)
             binding.rvMessageList.setBackgroundResource(recyclerViewBackground)
             binding.rvMessageList.setUseDivider(false)
             binding.vgTooltipBox.radius = resources.intToDp(19).toFloat()